feature gate random state
This is especially useful for no_std development, where we don't want to
introduce the getrandom dependency.

After this change, one can choose to disable random-state feature and
implement their own version of RandomState using whatever random source
they choose.

This also somewhat addresses #94 and avoids the need of using
compile-time-rng in restricted platforms.
