+++ tests/nopanic.rs
@@ -1,4 +1,4 @@
-use ahash::{AHasher, CallHasher, RandomState};
+use ahash::{AHasher, CallHasher};
 use std::hash::BuildHasher;
 
 #[macro_use]
@@ -48,9 +48,12 @@ fn hash_test_random_wrapper(num: i32, string: &str) {
     hash_test_specialize(num, string);
 }
 
+#[cfg(feature = "random-state")]
 #[inline(never)]
 #[no_panic]
 fn hash_test_random(num: i32, string: &str) -> (u64, u64) {
+    use ahash::RandomState;
+
     let build_hasher1 = RandomState::with_seeds(1, 2, 3, 4);
     let build_hasher2 = RandomState::with_seeds(1, 2, 3, 4);
     (
